requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
python-decouple==3.8
pytest
pytest-django
